update to glutin 0.31 and winit 0.30
These currently require raw-window-handle to be at version 0.5 to
interoperate, until glutin gets updated to use 0.6.

I had to remove support for headless rendering, but will try to add that
back at some point, if possible using EGL_KHR_surfaceless_context to
avoid the dependency on winit in that case.

The event loop now runs forever and gets started in main(), instead of
the previous way where it ran on demand from the GltfViewer.  This
accommodates for platforms which don’t let us run it on demand.